LESSON-6 fix extra folder

+++ otus_aibulat/movable/movable_adapter.py
@@ -1,8 +1,8 @@
 import math
 
-from otus_aibulat.lesson_6.movable.exceptions import GetLocationException, GetVelocityException, SetLocationException
-from otus_aibulat.lesson_6.types import SpaceObject, Vector
-from otus_aibulat.lesson_6.utils import raise_exception
+from otus_aibulat.movable.exceptions import GetLocationException, GetVelocityException, SetLocationException
+from otus_aibulat.types import SpaceObject, Vector
+from otus_aibulat.utils import raise_exception
         
 
 class MovableAdapter:
@@ -16,13 +16,13 @@ class MovableAdapter:
 
     @raise_exception(exception=GetVelocityException, message="Can't get velocity of movable")
     def get_velocity(self) -> Vector:
-            velocity: int = self.space_object["velocity"]
-            direction: int = self.space_object["direction"].direction
+        velocity: int = self.space_object["velocity"]
+        direction: float = self.space_object["direction"].direction
 
-            return Vector(
-                x=velocity*math.cos(direction),
-                y=velocity*math.sin(direction)
-            )
+        return Vector(
+            x=velocity*math.cos(direction),
+            y=velocity*math.sin(direction)
+        )
 
     @raise_exception(exception=SetLocationException, message="Can't set location for movable")
     def set_location(self, value: Vector) -> None:
